New effect of Lone Shadow
Similar to the previous description: After attacking an enemy, the player gains a 4-second "shadow buff," increasing speed by 10 points, 4 seconds of invisibility, and 5 seconds of darkness effect, while starting a 10-second cooldown. Nearby enemies are also inflicted with Slowness II for 4 seconds. During the "shadow buff," the total damage dealt by the player is accumulated and settled after 4 seconds. If the player switches to another weapon or if the Dahal is insufficient to maintain the "shadow buff," the damage settlement will occur prematurely.

When the "shadow buff" cooldown is 0, the player will be given particle effects while holding "Lone Shadow" for easier observation; when the "shadow buff" cooldown is greater than 1, there will be no particle effects.