A new spell that continuously converts nearby equipment into Chronotons and experience.
It exists as a "pet" in this map.

Players can select items of specific qualities for conversion in the book, and there is a "fool-proof" mechanism—the spell only converts items with "pickupdelay:0". This means if important equipment is accidentally dropped, it can be picked up immediately without being instantly cleared.

The positioning of this spell is relatively late-game. My idea is that after players reach Black Market level 5, they can synthesize it using the "Recycle" spell and some rune materials.

Because in the late game, the cooldown of the "Recycle" spell and the limitations on clearing items often can't keep up with the amount of items generated, I created an advanced version of the "Recycle" spell. This allows players to explore without worrying about the tedious and cumbersome process of handling excess equipment.

And it also has a hidden mechanism: after consuming a certain number of items, this pet will continuously spew out large amounts of diamonds, small coins, and big coins.